+++ optimal/src/OptimalError.ts
@@ -1,26 +1,23 @@
 import { ValidationError } from './ValidationError';
 
-export class OptimalError extends Error {
-	errors: ValidationError[] = [];
-
+export class OptimalError extends ValidationError {
 	file: string = '';
 
 	schema: string = '';
 
 	constructor() {
-		super('The following validations have failed:');
+		super('');
 
 		this.name = 'OptimalError';
 	}
 
-	addError(error: ValidationError) {
-		const prefix = error.path ? `Invalid field "${error.path}". ` : '';
-		const message = error.message
-			.split('\n')
-			.map((line) => (line.match(/^\s+-/g) ? `  ${prefix}${line}` : `  - ${prefix}${line}`))
-			.join('\n');
+	override addError(error: Error) {
+		const validError =
+			error instanceof ValidationError ? error : new ValidationError(error.message);
 
-		this.errors.push(error);
-		this.message += `\n${message}`;
+		this.errors.push(validError);
+
+		// Avoid indenting at this level
+		this.message = `${this.message}\n${error.message}`.trim();
 	}
 }