+++ optimal/src/ValidationError.ts
@@ -1,4 +1,8 @@
+import { pathKey } from './helpers';
+
 export class ValidationError extends Error {
+	errors: ValidationError[] = [];
+
 	path: string;
 
 	value: unknown;
@@ -9,5 +13,31 @@ export class ValidationError extends Error {
 		this.name = 'ValidationError';
 		this.path = path;
 		this.value = value;
+
+		if (path) {
+			const key = pathKey(path);
+
+			this.message = `${
+				key.includes('[') ? `Invalid member "${key}".` : `Invalid field "${key}".`
+			} ${this.message}`;
+		}
+	}
+
+	addError(error: Error) {
+		const validError =
+			error instanceof ValidationError ? error : new ValidationError(error.message);
+		const hasSameError = this.errors.some((e) => e.message === error.message);
+
+		if (hasSameError) {
+			return;
+		}
+
+		this.errors.push(validError);
+
+		this.message += '\n';
+		this.message += error.message
+			.split('\n')
+			.map((line) => (line.match(/^\s+-/) ? `  ${line}` : `  - ${line}`))
+			.join('\n');
 	}
 }