+++ optimal/src/criteria/shapes.ts
@@ -1,5 +1,6 @@
 import { invalid, invariant, isObject, isSchema, logUnknown } from '../helpers';
 import { Blueprint, Criteria, Schema, SchemaState, UnknownObject } from '../types';
+import { ValidationError } from '../ValidationError';
 
 /**
  * Require a shape to be an exact shape.
@@ -45,19 +46,36 @@ export function of<T extends object>(
 			const isPlainObject = value.constructor === Object;
 			const unknown: Partial<T> = isPlainObject ? { ...value } : {};
 			const shape: Partial<T> = {};
+			const collectionError = new ValidationError(
+				'The following validations have failed:',
+				path,
+				value,
+			);
 
 			Object.keys(schemas).forEach((prop) => {
 				const key = prop as keyof T;
 				const schema = schemas[key];
 
-				shape[key] = schema.validate(value[key], path ? `${path}.${key}` : String(key), {
-					...validateOptions,
-					currentObject: value as UnknownObject,
-				});
+				try {
+					shape[key] = schema.validate(value[key], path ? `${path}.${key}` : String(key), {
+						...validateOptions,
+						currentObject: value as UnknownObject,
+					});
+				} catch (error: unknown) {
+					if (error instanceof Error && validateOptions.collectErrors) {
+						collectionError.addError(error);
+					} else {
+						throw error;
+					}
+				}
 
 				// Delete the prop and mark it as known
 				delete unknown[key];
 			});
+
+			if (collectionError.errors.length > 0) {
+				throw collectionError;
+			}
 
 			// Handle unknown fields
 			if (state.metadata.exact) {