+++ optimal/CHANGELOG.md
@@ -25,13 +25,14 @@ changelog will use the new verbiage, but may affect previous APIs.
 #### 🚀 Updates
 
 - Chained methods are now immutable, returning new schema instances.
+- Errors are now collected and displayed as one, instead of failing on the first error.
 - Added `.notRequired()` method.
 - Added `date()` and `schema()` schemas.
 - Added `object().keysOf()` criteria method.
 - Added `createSchema()` function for creating custom schemas.
 - Added `createPredicate()` function for creating boolean return checkers.
 - Added custom messages to all schema methods through a `message` option.
-- Added lazy default value initializer callbacks to all methods.
+- Added lazy default value initializer callbacks to all schemas.
 
 #### 🛠 Internals